feat(ingress): new HTTP stage
* implement future internal package

* http ingress runner backbone

* implement http ingress env

* add correlation id propagation in processor stages

* add egress and link

* add http example

* http ingress/egress refactor

* add tls (http 2) config

* chore: add gauge metric

* cfirst metrics

* add p0 http ingress metrics

* update metrics-gen template

* clear generated metrics

* initial implementation of metrics-gen with attributes

* massive metrics improvements on http ingress stage

* update dashboards

* added tests and improved metrics-gen

+++ egress/quest_db.go
@@ -356,7 +356,7 @@ func (qw *questDBWorker) Deliver(ctx context.Context, msgIn *msg[*QuestDBMessage
 	span.SetAttributes(attribute.Int64("inserted_rows", int64(tmpInsRows)))
 
 	// Update metrics
-	qw.Env.Metrics.AddInsertedRows(uint(tmpInsRows))
+	qw.Env.Metrics.AddInsertedRows(tmpInsRows)
 
 	return nil
 }
